refactor: 세션 패키지 수정 및 캡슐화
