+++ tomcat/src/main/java/org/apache/coyote/http11/HttpRequest.java
@@ -1,7 +1,5 @@
 package org.apache.coyote.http11;
 
-import com.techcourse.model.Session;
-
 import java.util.Map;
 
 public class HttpRequest {
