+++ tomcat/src/main/java/org/apache/coyote/http11/Session.java
@@ -1,4 +1,4 @@
-package com.techcourse.model;
+package org.apache.coyote.http11;
 
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
